Adjust workflow to setup GitHub Issues labels
This is required to efficiently set up a new repo along with the issues
tracking labels that it requires.

It's set up to run each time there's a `push` action on the `ci` branch.

If I would like to add/edit labels using the GitHub UI afterwards, set
the `dry_run` key to `true`